1920 - Address PR review feedback: cast at true creation point, 2dp reconciliation rounding, trim comments
Moves each Float32/Enum cast to the column's actual point of creation across
_02_clean, _03_impute, _04_model, and _06_estimate_filled_posts instead of
batching them at the end of estimate_ind_cqc_filled_posts.py. Simplifies the
job-role reconciliation check to round the difference to 2dp and check it
falls between 0 and 1, replacing the relative/absolute tolerance pair. Trims
ticket-number references from comments per reviewer feedback.

+++ projects/_03_independent_cqc/_04_model/utils/model_utils.py
@@ -123,6 +123,6 @@ def add_predictions_into_df(
     run_id_value = f"v{model_version}_r{run_number}"
 
     return df.with_columns(
-        pl.Series(IndCQC.prediction, predictions),
+        pl.Series(IndCQC.prediction, predictions, dtype=pl.Float32),
         pl.lit(run_id_value).alias(IndCQC.prediction_run_id),
     )
